+++ trunk/lasers/src/edu/colorado/phet/lasers/model/mirror/BandPass.java
@@ -3,6 +3,11 @@
  * Package: edu.colorado.phet.lasers.model.mirror
  * Author: Another Guy
  * Date: Apr 2, 2003
+ * Latest Change:
+ *      $Author$
+ *      $Date$
+ *      $Name$
+ *      $Revision$
  */
 package edu.colorado.phet.lasers.model.mirror;
 
@@ -10,10 +15,10 @@ import edu.colorado.phet.lasers.model.photon.Photon;
 
 public class BandPass implements ReflectionStrategy {
 
-    private float cutoffLow;
-    private float cutoffHigh;
+    private double cutoffLow;
+    private double cutoffHigh;
 
-    public BandPass( float cutoffLow, float cutoffHigh ) {
+    public BandPass( double cutoffLow, double cutoffHigh ) {
         this.cutoffLow = cutoffLow;
         this.cutoffHigh = cutoffHigh;
     }
